Move PaymentDetails to own component and use in new journey
